+++ src/components/directives/safe-on/test/b-safe-on-dynamic-event-dummy/index.js
@@ -1,0 +1,2 @@
+package('b-safe-on-dynamic-event-dummy')
+	.extends('b-dummy');